Consistently use Long for all flags
Having a mix of Int and Long is a subtle source of bugs and doesn't
save any space - flags in symbols / modifiers are always Long.

Example for a source of bugs:

```
scala> val LateShift = 47
LateShift: Int = 47

scala> val DEFERRED = 1 << 4
DEFERRED: Int = 16

scala> DEFERRED << LateShift
res2: Int = 524288

scala> (DEFERRED: Long) << LateShift
res3: Long = 2251799813685248
```

+++ src/compiler/scala/tools/nsc/ast/parser/Parsers.scala
@@ -2168,7 +2168,7 @@ self =>
      *  }}}
      */
     def paramClauses(owner: Name, contextBounds: List[Tree], ofCaseClass: Boolean): List[List[ValDef]] = {
-      var implicitmod = 0
+      var implicitmod = 0L
       var caseParam = ofCaseClass
       def paramClause(): List[ValDef] = {
         if (in.token == RPAREN)
@@ -2226,7 +2226,7 @@ self =>
       }
     }
 
-    def param(owner: Name, implicitmod: Int, caseParam: Boolean): ValDef = {
+    def param(owner: Name, implicitmod: Long, caseParam: Boolean): ValDef = {
       val start = in.offset
       val annots = annotations(skipNewLines = false)
       var mods = Modifiers(Flags.PARAM)
@@ -2246,7 +2246,7 @@ self =>
       }
       val nameOffset = in.offset
       val name = ident()
-      var bynamemod = 0
+      var bynamemod = 0L
       val tpt =
         if ((settings.YmethodInfer && !owner.isTypeName) && in.token != COLON) {
           TypeTree()
@@ -2273,7 +2273,7 @@ self =>
           expr()
         } else EmptyTree
       atPos(start, if (name == nme.ERROR) start else nameOffset) {
-        ValDef((mods | implicitmod.toLong | bynamemod) withAnnotations annots, name.toTermName, tpt, default)
+        ValDef((mods | implicitmod | bynamemod) withAnnotations annots, name.toTermName, tpt, default)
       }
     }
 
@@ -3088,10 +3088,10 @@ self =>
     }
     */
 
-    def localDef(implicitMod: Int): List[Tree] = {
+    def localDef(implicitMod: Long): List[Tree] = {
       val annots = annotations(skipNewLines = true)
       val pos = in.offset
-      val mods = (localModifiers() | implicitMod.toLong) withAnnotations annots
+      val mods = (localModifiers() | implicitMod) withAnnotations annots
       val defs =
         if (!(mods hasFlag ~(Flags.IMPLICIT | Flags.LAZY))) defOrDcl(pos, mods)
         else List(tmplDef(pos, mods))
