+++ src/compiler/scala/tools/nsc/transform/CleanUp.scala
@@ -87,10 +87,10 @@ abstract class CleanUp extends Statics with Transform with ast.TreeDSL {
 
         def addStaticVariableToClass(forName: TermName, forType: Type, forInit: Tree, isFinal: Boolean): Symbol = {
           val flags = PRIVATE | STATIC | SYNTHETIC | (
-            if (isFinal) FINAL else 0
+            if (isFinal) FINAL else 0L
           )
 
-          val varSym = currentClass.newVariable(mkTerm("" + forName), ad.pos, flags.toLong) setInfoAndEnter forType
+          val varSym = currentClass.newVariable(mkTerm("" + forName), ad.pos, flags) setInfoAndEnter forType
           if (!isFinal)
             varSym.addAnnotation(VolatileAttr)
